fixed BUG-42908: sequences with a big start value fail
  Sequences with a start value bigger than max integer value(2147483647)
  were not supported throwing exception when updating database.

  Internally dbsm used int variables to store this value. It has now
  been changed to long as it is the type PostgreSQL returns from queries
  to pg_sequence.

+++ src/org/apache/ddlutils/model/Sequence.java
@@ -32,7 +32,7 @@ public class Sequence implements StructureObject, Cloneable {
   /** The name of the sequence, may be <code>null</code>. */
   private String _name;
   /** The start attribute of the sequence. */
-  private int _start;
+  private long _start;
   /** The increment attribute of the sequence. */
   private int _increment;
 
@@ -73,7 +73,7 @@ public class Sequence implements StructureObject, Cloneable {
    * 
    * @return The start attribute
    */
-  public int getStart() {
+  public long getStart() {
     return _start;
   }
 
@@ -83,7 +83,7 @@ public class Sequence implements StructureObject, Cloneable {
    * @param start
    *          The start attribute
    */
-  public void setStart(int start) {
+  public void setStart(long start) {
     _start = start;
   }
