+++ src/org/apache/ddlutils/platform/ModelLoaderBase.java
@@ -1,6 +1,6 @@
 /*
  ************************************************************************************
- * Copyright (C) 2001-2019 Openbravo S.L.U.
+ * Copyright (C) 2001-2020 Openbravo S.L.U.
  * Licensed under the Apache Software License version 2.0
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to  in writing,  software  distributed
@@ -624,7 +624,7 @@ public abstract class ModelLoaderBase implements ModelLoader {
         _log.debug("Sequence " + r.getString(1));
         Sequence s = new Sequence();
         s.setName(r.getString(1));
-        s.setStart(r.getInt(2));
+        s.setStart(r.getLong(2));
         s.setIncrement(r.getInt(3));
         return s;
       }
