+++ src/org/apache/ddlutils/platform/postgresql/PostgreSql10ModelLoader.java
@@ -1,6 +1,6 @@
 /*
  ************************************************************************************
- * Copyright (C) 2017 Openbravo S.L.U.
+ * Copyright (C) 2017-2020 Openbravo S.L.U.
  * Licensed under the Apache Software License version 2.0
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to  in writing,  software  distributed
@@ -46,7 +46,7 @@ public class PostgreSql10ModelLoader extends PostgreSqlModelLoader {
           readList(stmtCurrentSequence, new RowConstructor() {
             @Override
             public Object getRow(ResultSet seqDetails) throws SQLException {
-              sequence.setStart(seqDetails.getInt(1));
+              sequence.setStart(seqDetails.getLong(1));
               sequence.setIncrement(seqDetails.getInt(2));
               return sequence;
             }
