+++ src/org/apache/ddlutils/platform/postgresql/PostgreSqlModelLoader.java
@@ -1,6 +1,6 @@
 /*
  ************************************************************************************
- * Copyright (C) 2001-2019 Openbravo S.L.U.
+ * Copyright (C) 2001-2020 Openbravo S.L.U.
  * Licensed under the Apache Software License version 2.0
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to  in writing,  software  distributed
@@ -757,7 +757,7 @@ public class PostgreSqlModelLoader extends ModelLoaderBase {
           readList(stmtCurrentSequence, new RowConstructor() {
             @Override
             public Object getRow(ResultSet seqDetails) throws SQLException {
-              sequence.setStart(seqDetails.getInt(1));
+              sequence.setStart(seqDetails.getLong(1));
               sequence.setIncrement(seqDetails.getInt(2));
               return sequence;
             }
